+++ test/_hooks.js
@@ -1,6 +1,6 @@
 /* eslint-env node, mocha */
 import { drain as migratorDrain } from '../migrations/_migrator'
-import { drain } from '../src/DatabaseHandler'
+import { drain } from '../src/database'
 
 after(() => {
   migratorDrain()
